Add option to expose useCapture
This patch adds a new 'options' parameter to the 'on', 'off', and 'add'
methods, which is an object containing options. There is a single new option
available via this mechanism: 'useCapture'.

When 'on', 'off', or 'add' receive an options object with the 'useCapture' flag
set to true, listeners either added or removed to the DOM are done so with the
'useCapture' parameter of 'addEventListener' set to true. Previously, this
option was set to false for all listeners.

This resolves an issue with event bubbling in Mobile Safari, by giving us a way
to watch for these events during the Capture phase, instead of Bubble; because
some events, like 'input', are not propagating all the way to 'document' during
the Bubble stage.

Events like this that we want to watch during Capture, instead of Bubble,
should be configured like so:

	bean.on(document, 'input', '[type=range]', changeFn, {useCapture: true})

Note that when explicitly removing such listeners, you must also pass the option
to 'off':

	bean.off(document, 'input', changeFn, {useCapture: true})

The default in all cases is false, preserving backwards compatibility with the
current default.

References compose #260.